+++ src/zenml/cli/__init__.py
@@ -1105,6 +1105,6 @@ from zenml.cli.secret import *  # noqa
 from zenml.cli.server import *  # noqa
 from zenml.cli.stack import *  # noqa
 from zenml.cli.stack_components import *  # noqa
-# from zenml.cli.stack_recipes import *  # noqa
+from zenml.cli.stack_recipes import *  # noqa
 from zenml.cli.user_management import *  # noqa
 from zenml.cli.version import *  # noqa
